requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0
openai==1.3.0
anthropic==0.7.0
transformers==4.36.0
torch==2.1.0
ollama==0.1.7
groq==0.4.1
together==0.2.7
replicate==0.15.4
gitpython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
tree-sitter-typescript==0.20.2
tree-sitter-java==0.20.2
tree-sitter-cpp==0.20.0
tree-sitter-c==0.20.6
tree-sitter-rust==0.20.4
tree-sitter-go==0.20.0
psutil==5.9.6
watchdog==3.0.0
# File processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27
# Context management
tiktoken==0.5.2
sentence-transformers==2.2.2
numpy==1.24.3
